General code cleanup + various smaller fixes

+++ src/input/InputManager.cpp
@@ -678,9 +678,12 @@ EmulatedControllerPtr InputManager::delete_controller(size_t player_index, bool 
 		{
 			controller = {};
 
-			std::error_code ec{};
-			fs::remove(ActiveSettings::GetConfigPath("controllerProfiles/controller{}.xml", player_index), ec);
-			fs::remove(ActiveSettings::GetConfigPath("controllerProfiles/controller{}.txt", player_index), ec);
+			if(delete_profile)
+			{
+				std::error_code ec{};
+				fs::remove(ActiveSettings::GetConfigPath("controllerProfiles/controller{}.xml", player_index), ec);
+				fs::remove(ActiveSettings::GetConfigPath("controllerProfiles/controller{}.txt", player_index), ec);
+			}
 
 			return result;
 		}
@@ -840,7 +843,7 @@ std::optional<glm::ivec2> InputManager::get_left_down_mouse_info(bool* is_pad)
 		*is_pad = false;
 
 	{
-		std::shared_lock lock(m_main_mouse.m_mutex);
+		std::unique_lock lock(m_main_mouse.m_mutex);
 		if (std::exchange(m_main_mouse.left_down_toggle, false))
 			return m_main_mouse.position;
 
@@ -849,7 +852,7 @@ std::optional<glm::ivec2> InputManager::get_left_down_mouse_info(bool* is_pad)
 	}
 
 	{
-		std::shared_lock lock(m_main_touch.m_mutex);
+		std::unique_lock lock(m_main_touch.m_mutex);
 		if (std::exchange(m_main_touch.left_down_toggle, false))
 			return m_main_touch.position;
 
@@ -861,7 +864,7 @@ std::optional<glm::ivec2> InputManager::get_left_down_mouse_info(bool* is_pad)
 		*is_pad = true;
 
 	{
-		std::shared_lock lock(m_pad_mouse.m_mutex);
+		std::unique_lock lock(m_pad_mouse.m_mutex);
 		if (std::exchange(m_pad_mouse.left_down_toggle, false))
 			return m_pad_mouse.position;
 
@@ -870,7 +873,7 @@ std::optional<glm::ivec2> InputManager::get_left_down_mouse_info(bool* is_pad)
 	}
 
 	{
-		std::shared_lock lock(m_pad_touch.m_mutex);
+		std::unique_lock lock(m_pad_touch.m_mutex);
 		if (std::exchange(m_pad_touch.left_down_toggle, false))
 			return m_pad_touch.position;
 
@@ -887,7 +890,7 @@ std::optional<glm::ivec2> InputManager::get_right_down_mouse_info(bool* is_pad)
 		*is_pad = false;
 
 	{
-		std::shared_lock lock(m_main_mouse.m_mutex);
+		std::unique_lock lock(m_main_mouse.m_mutex);
 		if (std::exchange(m_main_mouse.right_down_toggle, false))
 			return m_main_mouse.position;
 
@@ -896,7 +899,7 @@ std::optional<glm::ivec2> InputManager::get_right_down_mouse_info(bool* is_pad)
 	}
 
 	{
-		std::shared_lock lock(m_main_touch.m_mutex);
+		std::unique_lock lock(m_main_touch.m_mutex);
 		if (std::exchange(m_main_touch.right_down_toggle, false))
 			return m_main_touch.position;
 
@@ -908,7 +911,7 @@ std::optional<glm::ivec2> InputManager::get_right_down_mouse_info(bool* is_pad)
 		*is_pad = true;
 
 	{
-		std::shared_lock lock(m_pad_mouse.m_mutex);
+		std::unique_lock lock(m_pad_mouse.m_mutex);
 		if (std::exchange(m_pad_mouse.right_down_toggle, false))
 			return m_pad_mouse.position;
 
@@ -917,7 +920,7 @@ std::optional<glm::ivec2> InputManager::get_right_down_mouse_info(bool* is_pad)
 	}
 
 	{
-		std::shared_lock lock(m_pad_touch.m_mutex);
+		std::unique_lock lock(m_pad_touch.m_mutex);
 		if (std::exchange(m_pad_touch.right_down_toggle, false))
 			return m_pad_touch.position;
